feat(51982): Inclui e altera informações cadastro de membros da associação
