+++ src/componentes/escolas/Associacao/associacao.scss
@@ -51,3 +51,7 @@
 .link-green:hover{
   cursor: pointer;
 }
+.switch-status-presidente-checked{
+  background-color: #307D83 !important;
+  color: #fff !important;
+}
